refactor home index
